+++ src/Products/GenericSetup/tool.py
@@ -27,6 +27,7 @@ from operator import itemgetter
 import six
 
 from AccessControl.class_init import InitializeClass
+from AccessControl.Permissions import view
 from AccessControl.SecurityInfo import ClassSecurityInfo
 from Acquisition import aq_base
 from OFS.Folder import Folder
@@ -1556,6 +1557,11 @@ class SetupTool(Folder):
                     content_type='text/plain')
 
         self._setObject(name, file)
+
+        # Tighten the View permission on the log file. Only the owner and
+        # Manager users may view the log.
+        file_ob = self._getOb(name)
+        file_ob.manage_permission(view, ('Manager', 'Owner'), 0)
 
 
 InitializeClass(SetupTool)